Cast metric to the same type as moving mean

+++ larq/metrics.py
@@ -48,7 +48,7 @@ class MeanChangedValues(tf.keras.metrics.Mean):
             return self._previous_values.assign(values)
         changed_values = tf.math.count_nonzero(tf.equal(self._previous_values, values))
         metric_update_op = super(MeanChangedValues, self).update_state(
-            1 - (tf.cast(changed_values, tf.float32) / self._size)
+            1 - (tf.cast(changed_values, self.dtype) / self._size)
         )
         with tf.control_dependencies([metric_update_op]):
             return self._previous_values.assign(values)
